+++ tests/secret-exec-migrate.zsh
@@ -16,7 +16,7 @@ trap 'rm -rf -- "$test_dir"' EXIT
 
 # zsh skips EXIT traps when errexit fires inside a function.
 TRAPZERR() {
-  if [[ -o errexit ]] && (( ${#funcstack} > 1 )); then
+  if [[ -o errexit ]] && (( ZSH_SUBSHELL == 0 && ${#funcstack} > 1 )); then
     rm -rf -- "$test_dir"
   fi
 }
